Alter Make Command to provide docblocks that help with IDE Autocomplete and also a flag to be able to set the model
The main part of this work was adding the docblocks to help IDEs with
autocompletion, which involved working out the Model class from the
factory name.

However, as there is also the option of changing the model name for the
Poser factory, a `-model` flag was added that will allow for specifying
the model at the point of creation.

The change to `GeneratorCommand` was due to it providing
`$this->qualifyClass()` that will work out the namespace in the same
way that Laravels make commands do